+++ src/components/VideoProvider/index.tsx
@@ -3,6 +3,8 @@
  *
  * Add HuddleProvider inside the VideoContext.
  * Add WidgetProvider inside the VideoContext.
+ * Add RoomStateProvider.
+ * Add ParticipantMetaProvider.
  * `localTracks` context property can now include LocalDataTracks.
  */
 
@@ -21,9 +23,11 @@ import { RoomStateProvider } from '../../withComponents/RoomState/RoomStateProvi
 import { HuddleProvider } from '../../withComponents/HuddleProvider/HuddleProvider';
 import { WidgetProvider } from '../../withComponents/WidgetProvider/WidgetProvider';
 import { RoomMetaProvider } from '../../withComponents/RoomMetaProvider/RoomMetaProvider';
+import { ParticipantMetaProvider } from '../../withComponents/ParticipantMetaProvider/ParticipantMetaProvider';
 import huddles from '../../withComponents/HuddleProvider/huddleReducer';
 import widgets from '../../withComponents/WidgetProvider/widgetReducer';
 import properties from '../../withComponents/RoomMetaProvider/roomMetaReducer';
+import participantMeta from '../../withComponents/ParticipantMetaProvider/participantMetaReducer';
 
 /*
  *  The hooks used by the VideoProvider component are different than the hooks found in the 'hooks/' directory. The hooks
@@ -77,13 +81,15 @@ export function VideoProvider({ options, children, onError = () => {}, onDisconn
         connect,
       }}
     >
-      <RoomStateProvider reducers={{ huddles, widgets, properties }}>
+      <RoomStateProvider reducers={{ huddles, widgets, properties, participantMeta }}>
         <RoomMetaProvider>
-          <HuddleProvider>
-            <WidgetProvider>
-              <SelectedParticipantProvider room={room}>{children}</SelectedParticipantProvider>
-            </WidgetProvider>
-          </HuddleProvider>
+          <ParticipantMetaProvider>
+            <HuddleProvider>
+              <WidgetProvider>
+                <SelectedParticipantProvider room={room}>{children}</SelectedParticipantProvider>
+              </WidgetProvider>
+            </HuddleProvider>
+          </ParticipantMetaProvider>
         </RoomMetaProvider>
       </RoomStateProvider>
     </VideoContext.Provider>